# LinkedList optimizations