### Some benchmarks I've made
<details><summary>Before</summary>
<p>

```julia
julia> using DataStructures, BenchmarkTools

julia> x = list(1:1e7...);

julia> @benchmark copy(x)
BenchmarkTools.Trial:
  memory estimate:  1.79 GiB
  allocs estimate:  80000002
  --------------
  minimum time:     3.447 s (29.88% GC)
  median time:      3.759 s (39.81% GC)
  mean time:        3.759 s (39.81% GC)
  maximum time:     4.071 s (48.21% GC)
  --------------
  samples:          2
  evals/sample:     1



julia> @benchmark map(x->x, x)
BenchmarkTools.Trial: 
  memory estimate:  1.79 GiB
  allocs estimate:  79999999
  --------------
  minimum time:     3.670 s (28.29% GC)
  median time:      4.006 s (37.79% GC)
  mean time:        4.006 s (37.79% GC)
  maximum time:     4.341 s (45.83% GC)
  --------------
  samples:          2
  evals/sample:     1

julia> @benchmark filter(x->x%2==0, x)
BenchmarkTools.Trial: 
  memory estimate:  1.19 GiB
  allocs estimate:  55000002
  --------------
  minimum time:     2.946 s (28.14% GC)
  median time:      3.218 s (34.64% GC)
  mean time:        3.218 s (34.64% GC)
  maximum time:     3.489 s (40.13% GC)
  --------------
  samples:          2
  evals/sample:     1

```

</p>
</details>


<details><summary>After</summary>
<p>

```julia
julia> using DataStructures, BenchmarkTools

julia> x = list(1:1e7...);

julia> @benchmark copy(x)
BenchmarkTools.Trial: 
  memory estimate:  915.53 MiB
  allocs estimate:  39999998
  --------------
  minimum time:     1.710 s (20.68% GC)
  median time:      2.055 s (31.53% GC)
  mean time:        2.040 s (33.19% GC)
  maximum time:     2.354 s (43.72% GC)
  --------------
  samples:          3
  evals/sample:     1

julia> @benchmark map(x->x, x)
BenchmarkTools.Trial: 
  memory estimate:  915.53 MiB
  allocs estimate:  39999998
  --------------
  minimum time:     1.708 s (21.05% GC)
  median time:      2.034 s (31.46% GC)
  mean time:        2.029 s (33.26% GC)
  maximum time:     2.345 s (43.70% GC)
  --------------
  samples:          3
  evals/sample:     1

julia> @benchmark filter(x->x%2==0, x)
BenchmarkTools.Trial: 
  memory estimate:  762.94 MiB
  allocs estimate:  35000002
  --------------
  minimum time:     1.715 s (19.23% GC)
  median time:      1.934 s (24.52% GC)
  mean time:        1.877 s (23.94% GC)
  maximum time:     1.981 s (27.46% GC)
  --------------
  samples:          3
  evals/sample:     1
```

</p>
</details>