@oxinabox 2020-10-15
While these are very nice and cool,
we are planning on removing LinkedList from the package; because they better belong in JuliaCollections/FunctionalCollections.jl
There is a similar linked list in that package.
You can investigate openneing a similar PR there